requirements.txt
discord.py
imageio
Pillow
